# promises for chops, active bindings for slices