--- a/R/slice.R
+++ b/R/slice.R
@@ -253,6 +253,8 @@ slice_rows <- function(.data, ...) {
   }
 
   mask <- DataMask$new(.data, caller_env())
+  on.exit(mask$forget("slice"), add = TRUE)
+
   rows <- mask$get_rows()
 
   quo <- quo(c(!!!dots))