--- a/src/filter.cpp
+++ b/src/filter.cpp
@@ -137,7 +137,7 @@ SEXP dplyr_mask_eval_all_filter(SEXP quos, SEXP env_private, SEXP s_n, SEXP env_
     SEXP rows_i = VECTOR_ELT(rows, i);
     R_xlen_t n_i = XLENGTH(rows_i);
 
-    SEXP result_i = PROTECT(eval_filter_one(quos, VECTOR_ELT(masks, i), caller, n_i, env_filter));
+    SEXP result_i = PROTECT(eval_filter_one(quos, mask, caller, n_i, env_filter));
 
     int* p_rows_i = INTEGER(rows_i);
     int* p_result_i = LOGICAL(result_i);