This builds on top of #5520 but instead of creating one rlang data mask per group with promises, this has only one data mask, but holding active bindings. 

The active binding still uses the lazy chops which still live in an environment full of promises. 

For example with `data.frame(x = 1:2, y = 3:4) %>% group_by(x)` we create: 

 - the lazy chops: an environment with promises of chops for `x` and `y`. The parent of this env has the `.current_group` variable that needs to be changed as the evaluation moves from one group to the other. 
 - the rlang data mask, with bindings like `function() .subset2(x, .current_group)`. The binding function's enclosure is the lazy chops env

``` r
library(dplyr)
set.seed(123)

group_size <- 5L
n_groups <- 1e5
size <- n_groups * group_size

g <- rep(seq_len(n_groups), each = group_size)
x <- rnorm(size)

df <- tibble(x = x, g = g)

gdf <- group_by(df, g)
```

This PR: 

```r
bench::mark(
  sum = summarise(gdf, x = sum(x), .groups = "drop_last")
)[1:8]
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 1 x 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 sum           141ms    166ms      5.75    7.41MB     15.8
```

#5520 (with promises and one mask per group): 

``` r
bench::mark(
  sum = summarise(gdf, x = sum(x), .groups = "drop_last")
)[1:8]
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 1 x 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 sum           1.59s    1.59s     0.628     137MB     3.77
```

master: 

``` r
bench::mark(
  sum = summarise(gdf, x = sum(x), .groups = "drop_last")
)[1:8]
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 1 x 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 sum           116ms    128ms      7.81    7.49MB     9.77
```

<sup>Created on 2020-10-27 by the [reprex package](https://reprex.tidyverse.org) (v0.3.0.9001)</sup>

Performance feels similar to master on the "many groups" case. Presumably using an active binding (instead of the adhoc system we had to update the mask based on what has been used/resolved) is slightly more expensive.