@romainfrancois 2020-11-02
merging this into #5520